mvp_0.5/requirements.txt
pyyaml>=6.0
urdfpy>=0.0.22
numpy>=1.21.0
# CI Pipeline Dependencies
gitpython>=3.1.0
tree-sitter>=0.20.0
sdformat>=13.0.0
mjcf>=2.0.0
SALib>=1.4.0
mujoco>=2.3.0
pybullet>=3.2.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
# ROS 2 parsing
lxml>=4.9.0
# Additional utilities
pathlib2>=2.3.0
tqdm>=4.64.0
